+++ SistemaAcademia/SistemaAcademia/Models/Academia.cs
@@ -1,10 +1,10 @@
 namespace SistemaAcademia.Models;
 
-public class Academia(string nome, string cnpj, string email, int idEndereco)
+public class Academia(string nome, string cnpj, string email, long idEndereco)
 {
-    public int Id { get; init; }
+    public long Id { get; init; }
     public string Nome { get; } = nome;
     public string CNPJ { get; } = cnpj;
     public string Email { get; } = email;
-    public int IdEndereco { get; } = idEndereco;
+    public long IdEndereco { get; } = idEndereco;
 }
