+++ SistemaAcademia/SistemaAcademia/Models/Cliente.cs
@@ -1,8 +1,8 @@
 namespace SistemaAcademia.Models;
 
-public class Cliente(string nome, string cpf, string email, string senha, DateTime dataNascimento, int idUsuario)
+public class Cliente(string nome, string cpf, string email, string senha, DateTime dataNascimento, long idUsuario)
     : Pessoa(nome, cpf, email, senha, dataNascimento)
 {
-    public int IdAcademia { get; init; }
-    public int IdUsuario { get; } = idUsuario;
+    public long IdAcademia { get; init; }
+    public long IdUsuario { get; } = idUsuario;
 }
