+++ SistemaAcademia/SistemaAcademia/Models/Modalidade.cs
@@ -1,9 +1,9 @@
 namespace SistemaAcademia.Models;
 
-public class Modalidade(string nome, decimal valorModalidade, int idAcademia)
+public class Modalidade(string nome, decimal valorModalidade, long idAcademia)
 {
-    public int     Id              { get; init; }
+    public long    Id              { get; init; }
     public string  Nome            { get; } = nome;
     public decimal ValorModalidade { get; } = valorModalidade;
-    public int     IdAcademia      { get; } = idAcademia;
+    public long    IdAcademia      { get; } = idAcademia;
 }
