+++ SistemaAcademia/SistemaAcademia/Models/Pagamento.cs
@@ -2,10 +2,10 @@ using SistemaAcademia.Enum;
 
 namespace SistemaAcademia.Models;
 
-public class Pagamento(int idMatricula, decimal valorPrevisto, DateTime dataVencimento)
+public class Pagamento(long idMatricula, decimal valorPrevisto, DateTime dataVencimento)
 {
-    public int Id { get; init; }
-    public int IdMatricula { get; } = idMatricula;
+    public long Id { get; init; }
+    public long IdMatricula { get; } = idMatricula;
     public decimal ValorPrevisto { get; } = valorPrevisto;
     public decimal? ValorPago { get; set; }
     public DateTime DataVencimento { get; } = dataVencimento;
