+++ SistemaAcademia/SistemaAcademia/Models/PagamentoViewModel.cs
@@ -2,8 +2,8 @@ namespace SistemaAcademia.Models;
 
 public class PagamentoListaViewModel
 {
-    public int      Id               { get; set; }
-    public int      IdMatricula      { get; set; }
+    public long     Id               { get; set; }
+    public long     IdMatricula      { get; set; }
     public string   NomeModalidade   { get; set; } = string.Empty;
     public string   NomePlano        { get; set; } = string.Empty;
     public decimal  ValorEsperado    { get; set; }
@@ -16,14 +16,14 @@ public class PagamentoListaViewModel
 
 public class RegistrarPagamentoViewModel
 {
-    public int     IdPagamento       { get; set; }
-    public int     IdMetodoPagamento { get; set; }
+    public long    IdPagamento       { get; set; }
+    public long    IdMetodoPagamento { get; set; }
     public decimal ValorPago         { get; set; }
-    public int     IdCliente         { get; set; }
+    public long    IdCliente         { get; set; }
 }
 
 public class MetodoPagamentoViewModel
 {
-    public int    Id   { get; set; }
+    public long   Id   { get; set; }
     public string Nome { get; set; } = string.Empty;
 }
